feat(Process): Remove Processing algorithms from Manager.py to Process.py.
