+++ .idea/Loop System.iml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <module type="PYTHON_MODULE" version="4">
   <component name="NewModuleRootManager">
+    <content url="file://$MODULE_DIR$/../Python-Plot" />
+    <content url="file://$MODULE_DIR$/../Study1/Physiological_Data" />
     <content url="file://$MODULE_DIR$" />
     <orderEntry type="jdk" jdkName="Python 3.10 (Loop System)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
